[FE] REFACTOR: chat refactor #121
