+++ frontend/src/components/modules/Chat/Chat.tsx
@@ -8,6 +8,7 @@ import { IMessage } from "../../atoms/ChatList/ChatList";
 import { useRecoilValue } from "recoil";
 import blockUsersState from "../../../state/BlockUsersState";
 import { IUser } from "../Pagination/Pagination";
+import ChatInput from "../ChatInput";
 
 interface Props {
   width: string;
@@ -37,69 +38,14 @@ const ChatBoard = styled(Board).attrs({
 })``;
 
 const Chat = ({ ...rest }: Props) => {
-  const [list, setList] = useState<IMessage[]>([]);
-  const [msg, setMsg] = useState<string>();
-  const blockUsers = useRecoilValue(blockUsersState);
   const socket = useContext(SocketContext);
-
-  useEffect(() => {
-    // const joinRoom = (newMsg: IMessage) => {
-    //   setList((prevList) => [...prevList, newMsg]);
-    // };
-    const newMessage = (newMsg: IMessage) => {
-      for (const user of blockUsers) {
-        if (user.userId == newMsg.fromId) return;
-      }
-      setList((prevList) => [...prevList, newMsg]);
-    };
-    if (socket) {
-      socket.socket.off("message");
-      socket.socket.off("whisper");
-      socket.socket.off("systemMsg");
-      socket.socket.on("message", newMessage);
-      socket.socket.on("whisper", newMessage);
-      socket.socket.on("systemMsg", newMessage);
-    }
-  }, [blockUsers]);
-
-  const enterkey = (event: KeyboardEvent<HTMLInputElement>) => {
-    if (event.key !== "Enter" || !msg || msg === "" || !socket) return;
-
-    if (msg.split(" ")[0] === "/w") {
-      socket.socket.emit("whisper", {
-        fromName: socket.userName,
-        toName: msg.split(" ")[1],
-        msg: msg.substring(msg.split(" ")[1].length + 4),
-      });
-    } else {
-      console.log(msg);
-      socket.socket.emit("message", {
-        roomId: 0,
-        userId: socket.userId,
-        userName: socket.userName,
-        msg: socket.userName + ": " + msg,
-      });
-    }
-    setMsg("");
-  };
 
   return (
     <SocketContext.Provider value={socket}>
       <ChatStyled {...rest}>
         <ChatBoard>
-          <ChatList list={list} height="87%" width="100%" />
-          <Input
-            width="100%"
-            height="10%"
-            borderRadius
-            padding="1rem"
-            fontSize="1.5rem"
-            onKeyPress={enterkey}
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-              setMsg(e.target.value)
-            }
-            value={msg ? msg : ""}
-          />
+          <ChatList height="87%" width="100%" />
+          <ChatInput />
         </ChatBoard>
       </ChatStyled>
     </SocketContext.Provider>
