+++ frontend/src/components/modules/ChatInput/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ChatInput";
